pyproject.toml
[project]
name = "duality"
version = "0.1.0"
description = "Semantic Scene Segmentation for Duality AI Offroad Autonomy Challenge"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "transformers>=4.35.0",
    "segmentation-models-pytorch>=0.3.3",
    "albumentations>=1.3.0",
    "opencv-contrib-python>=4.8.0",
    "tqdm>=4.66.0",
    "matplotlib>=3.8.0",
    "pillow>=10.0.0",
    "numpy>=1.24.0",
]
